// src/services/types.ts
export type Show = {
  id: number;
  url: string;
  name: string;
  type: string;
  language: string;
  genres: string[];
  status: string;
  runtime: number;
  averageRuntime: number;
  premiered: string;
  ended: string;
  officialSite: string;
  schedule: {
    time: string;
    days: string[];
  };
  rating: {
    average: number;
  };
  image: {
    medium: string;
    original: string;
  };
  summary: string;
  updated: number;
  _links: {
    self: {
      href: string;
    };
    previousepisode: {
      href: string;
    };
  };
  _embedded?: {
    previousepisode: Episode;
  };
};

export type Season = {
  id: number;
  url: string;
  number: number;
  name: '';
  episodeOrder: number;
  premiereDate: string;
  endDate: string;
  image: {
    medium: string;
    original: string;
  };
  summary: string;
  _links: {
    self: {
      href: string;
    };
  };
};

export type Episode = {
  id: number;
  url: string;
  name: string;
  season: number;
  number: number;
  type: string;
  airdate: string;
  airtime: string;
  airstamp: string;
  runtime: number;
  rating: {
    average: number;
  };
  image: {
    medium: string;
    original: string;
  };
  summary: string;
  _links: {
    self: {
      href: string;
    };
  };
};

export type ShowCategory = {
  genre: string;
  show: Show;
};

export type SearchShow = {
  score: number;
  show: Show;
};

export type EpisodesBySeason = Season & {
  _embedded?: {
    episodes: Episode[];
  };
};